requirements.txt
requests==2.31.0
python-dotenv==1.0.0
opencv-python==4.9.0.80
opencv-contrib-python==4.9.0.80
numpy==1.26.4